// src/generated/client/models/HomeAddress.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Megasense data portal API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HomeAddress
 */
export interface HomeAddress {
    /**
     * 
     * @type {string}
     * @memberof HomeAddress
     */
    streetAddress: string;
    /**
     * 
     * @type {string}
     * @memberof HomeAddress
     */
    postalCode: string;
    /**
     * 
     * @type {string}
     * @memberof HomeAddress
     */
    city: string;
    /**
     * 
     * @type {string}
     * @memberof HomeAddress
     */
    country: string;
}

export function HomeAddressFromJSON(json: any): HomeAddress {
    return HomeAddressFromJSONTyped(json, false);
}

export function HomeAddressFromJSONTyped(json: any, ignoreDiscriminator: boolean): HomeAddress {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'streetAddress': json['streetAddress'],
        'postalCode': json['postalCode'],
        'city': json['city'],
        'country': json['country'],
    };
}

export function HomeAddressToJSON(value?: HomeAddress | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'streetAddress': value.streetAddress,
        'postalCode': value.postalCode,
        'city': value.city,
        'country': value.country,
    };
}